[Improvement][AJ] Used action sheets for modals for smoother and less buggier animation/experience
